move search delegate to different file
