+++ lib/presentation/screens/new_filter/widgets/lecturer_filter/widgets/search/new_lecturer_search_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:silvertimetable/data/models/lecturer/lecturer_base.dart';
 import 'package:silvertimetable/logic/schedule_manager/schedule_manager_bloc.dart';
 import 'package:silvertimetable/presentation/screens/new_filter/widgets/lecturer_filter/cubits/lecturer_picked/lecturer_picked_cubit.dart';
-import 'package:silvertimetable/presentation/screens/new_filter/widgets/lecturer_filter/widgets/search/lecturer_search_tile.dart';
+import 'package:silvertimetable/presentation/screens/new_filter/widgets/lecturer_filter/widgets/search/new_lecturer_search_delegate.dart';
 
 class NewLecturerSearchButton extends StatefulWidget {
   const NewLecturerSearchButton({super.key});
@@ -16,8 +16,6 @@ class NewLecturerSearchButton extends StatefulWidget {
 class _NewLecturerSearchButtonState extends State<NewLecturerSearchButton> {
   @override
   Widget build(BuildContext context) {
-    // TODO: clean the code and fix showing suggestions and results when searching
-
     return BlocBuilder<ScheduleManagerBloc, ScheduleManagerState>(
       builder: (context, state) {
         return IconButton(
@@ -35,79 +33,6 @@ class _NewLecturerSearchButtonState extends State<NewLecturerSearchButton> {
           icon: const Icon(Icons.search),
         );
       },
-    );
-  }
-}
-
-class NewLecturerSearchDelegate extends SearchDelegate {
-  NewLecturerSearchDelegate({
-    required this.lecturers,
-    required this.bloc,
-  });
-
-  List<LecturerBase> lecturers;
-  LecturerPickedCubit bloc;
-
-  String lecturerInfoString(LecturerBase lecturer) =>
-      '${lecturer.academicDegree} ${lecturer.firstName} ${lecturer.surname} ${lecturer.email}';
-
-  @override
-  List<Widget>? buildActions(BuildContext context) {
-    return [
-      IconButton(
-        onPressed: () {
-          query = '';
-        },
-        icon: const Icon(Icons.clear),
-      ),
-    ];
-  }
-
-  @override
-  Widget? buildLeading(BuildContext context) {
-    return IconButton(
-      onPressed: () {
-        close(context, null);
-      },
-      icon: const Icon(Icons.arrow_back),
-    );
-  }
-
-  @override
-  Widget buildResults(BuildContext context) {
-    final List<LecturerBase> matched = [];
-    for (final lecturer in lecturers) {
-      if (lecturerInfoString(lecturer)
-          .toLowerCase()
-          .contains(query.toLowerCase())) {
-        matched.add(lecturer);
-      }
-    }
-    return ListView.builder(
-      itemCount: matched.length,
-      itemBuilder: (context, index) => BlocProvider.value(
-        value: bloc,
-        child: LecturerSearchTile(lecturer: matched[index]),
-      ),
-    );
-  }
-
-  @override
-  Widget buildSuggestions(BuildContext context) {
-    final List<LecturerBase> matched = [];
-    for (final lecturer in lecturers) {
-      if (lecturerInfoString(lecturer)
-          .toLowerCase()
-          .contains(query.toLowerCase())) {
-        matched.add(lecturer);
-      }
-    }
-    return ListView.builder(
-      itemCount: matched.length,
-      itemBuilder: (context, index) => BlocProvider.value(
-        value: bloc,
-        child: LecturerSearchTile(lecturer: matched[index]),
-      ),
     );
   }
 }
